Extract some test env data to TestEnv classes
